data-service/requirements.txt
bcrypt==4.2.0
grpcio==1.67.0
grpcio-tools==1.66.2
protobuf==5.28.3
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.3
python-dotenv==1.0.1
setuptools==75.2.0
typing_extensions==4.12.2
